# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEO Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /case-study:
    get:
      operationId: getCaseStudy
      tags: [caseStudies]
      summary: A company's case study, revision history, job state, and new-activity counts
      parameters:
        - name: companyId
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Case study overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudyOverview"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /case-study/context:
    put:
      operationId: saveCaseStudyContext
      tags: [caseStudies]
      summary: Save the authored starting position, strategic focus, and notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyContextInput"
      responses:
        "200":
          description: Saved case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudyRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /case-study/refresh:
    post:
      operationId: refreshCaseStudy
      tags: [caseStudies]
      summary: Start a background narrative refresh (poll /case-study/jobs)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyRefreshInput"
      responses:
        "202":
          description: Refresh job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudyJob"
        "409":
          description: A refresh is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "503":
          description: Generation unavailable (no model key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /case-study/jobs:
    get:
      operationId: listCaseStudyJobs
      tags: [caseStudies]
      summary: Recent case-study refresh jobs for a company, newest first
      parameters:
        - name: companyId
          in: query
          required: true
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStudyJob"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /case-study/revisions/{id}:
    patch:
      operationId: updateCaseStudyRevision
      tags: [caseStudies]
      summary: Edit the wording of a revision or approve/unapprove it
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyRevisionUpdate"
      responses:
        "200":
          description: Updated revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudyRevision"
        "404":
          description: Revision not found for this company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
...
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    CaseStudyRecord:
      type: object
      properties:
        id: { type: integer }
        companyId: { type: integer }
        startingPosition: { type: ["string", "null"] }
        strategicFocus: { type: ["string", "null"] }
        contextNotes: { type: ["string", "null"] }
        contextUpdatedAt: { type: ["string", "null"] }
        createdAt: { type: string }
      required:
        [id, companyId, startingPosition, strategicFocus, contextNotes, contextUpdatedAt, createdAt]
    CaseStudyContextInput:
      type: object
      properties:
        companyId: { type: integer }
        startingPosition: { type: ["string", "null"], maxLength: 4000 }
        strategicFocus: { type: ["string", "null"], maxLength: 4000 }
        contextNotes: { type: ["string", "null"], maxLength: 4000 }
      required: [companyId]
    CaseStudyMetricWindow:
      type: object
      properties:
        start: { type: string }
        end: { type: string }
        runs: { type: integer }
        visibilityPct: { type: ["number", "null"] }
        avgBrandPosition: { type: ["number", "null"] }
        verifiedCitations: { type: integer }
        ownedCitations: { type: integer }
      required:
        [start, end, runs, visibilityPct, avgBrandPosition, verifiedCitations, ownedCitations]
    CaseStudyEvidenceSummary:
      type: object
      properties:
        totalRuns: { type: integer }
        activePrompts: { type: integer }
        firstRunAt: { type: ["string", "null"] }
        lastRunAt: { type: ["string", "null"] }
        earliest:
          oneOf:
            - $ref: "#/components/schemas/CaseStudyMetricWindow"
            - type: "null"
        latest:
          oneOf:
            - $ref: "#/components/schemas/CaseStudyMetricWindow"
            - type: "null"
        actionsLive: { type: integer }
        actionsInProgress: { type: integer }
        measurementCount: { type: integer }
        gaps:
          type: array
          items: { type: string }
      required:
        [totalRuns, activePrompts, firstRunAt, lastRunAt, earliest, latest, actionsLive, actionsInProgress, measurementCount, gaps]
    CaseStudyRevision:
      type: object
      properties:
        id: { type: integer }
        caseStudyId: { type: integer }
        companyId: { type: integer }
        revisionNumber: { type: integer }
        kind: { type: string, enum: [baseline, refresh] }
        narrative: { type: string }
        editedNarrative: { type: ["string", "null"] }
        approvedAt: { type: ["string", "null"] }
        periodStart: { type: ["string", "null"] }
        periodEnd: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        createdAt: { type: string }
        evidence:
          $ref: "#/components/schemas/CaseStudyEvidenceSummary"
      required:
        [id, caseStudyId, companyId, revisionNumber, kind, narrative, editedNarrative, approvedAt, periodStart, periodEnd, model, createdAt, evidence]
    CaseStudyRefreshInput:
      type: object
      properties:
        companyId: { type: integer }
      required: [companyId]
    CaseStudyRevisionUpdate:
      type: object
      properties:
        companyId: { type: integer }
        editedNarrative: { type: ["string", "null"], maxLength: 20000 }
        approved: { type: boolean }
      required: [companyId]
    CaseStudyJob:
      type: object
      properties:
        id: { type: integer }
        companyId: { type: integer }
        status: { type: string, enum: [running, completed, failed] }
        stage: { type: string }
        error: { type: ["string", "null"] }
        revisionId: { type: ["integer", "null"] }
        createdAt: { type: string }
        finishedAt: { type: ["string", "null"] }
      required:
        [id, companyId, status, stage, error, revisionId, createdAt, finishedAt]
    CaseStudyNewActivity:
      type: object
      properties:
        runs: { type: integer }
        actions: { type: integer }
        measurements: { type: integer }
      required: [runs, actions, measurements]
    CaseStudyOverview:
      type: object
      properties:
        caseStudy:
          $ref: "#/components/schemas/CaseStudyRecord"
        revisions:
          type: array
          items:
            $ref: "#/components/schemas/CaseStudyRevision"
        activeJob:
          oneOf:
            - $ref: "#/components/schemas/CaseStudyJob"
            - type: "null"
        lastJob:
          oneOf:
            - $ref: "#/components/schemas/CaseStudyJob"
            - type: "null"
        newActivity:
          oneOf:
            - $ref: "#/components/schemas/CaseStudyNewActivity"
            - type: "null"
        generationAvailable: { type: boolean }
      required:
        [caseStudy, revisions, activeJob, lastJob, newActivity, generationAvailable]
...